refactor(web): hooks 위치와 query state 전이 네이밍 정리
